requirements.txt
requests
gitpython
pyyaml
scholarly
